SITL: estimate MAVLink gimbal state from requested telemetry
Request vehicle position and attitude over the gimbal link, predict position from received velocity, and combine received attitude with the three simulated encoders. Stop targeting and retry requests when required telemetry expires.

Keep command and feedback roll/pitch horizon-referenced, honor explicit yaw frames, and avoid vehicle-rate feedforward for earth-fixed commands. Unit tests cover freshness, prediction, encoder composition and command/status reference frames.

+++ libraries/SITL/SIM_AVT_CM62.h
@@ -36,7 +36,7 @@ private:
     const char *get_model_name()       const override { return "SIM_AVTA"; }
     // firmware_version encoding: major=1 | (minor=2)<<8 | (patch=3)<<16
     uint32_t    get_firmware_version() const override { return (3U << 16) | (2U << 8) | 1U; }
-    uint16_t    get_cap_flags()        const override {
+    uint32_t    get_cap_flags()        const override {
         return GIMBAL_DEVICE_CAP_FLAGS_HAS_PITCH_AXIS |
                GIMBAL_DEVICE_CAP_FLAGS_HAS_YAW_AXIS;
     }
